Rename gateway.py to target.py
The next patches move the gateway management code into a new class.
The target code has already taken the filename gateway.py, so this patch
just renames it to target.py.
